# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /admin/registration/{type}/{id}/reject:
    put:
      tags:
      - Registration
      operationId: rejectRegistrationRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum:
          - ADMIN
          - CLIENT
          - FREELANCER
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /admin/registration/{type}/{id}/approve:
    put:
      tags:
      - Registration
      operationId: approveRegistrationRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum:
          - ADMIN
          - CLIENT
          - FREELANCER
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
...
  /admin/freelancer-registrations:
    get:
      tags:
      - Registration
      operationId: getFreelancersRegistrationRequests
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationRequestDto'
  /admin/client-registrations:
    get:
      tags:
      - Registration
      operationId: getClientsRegistrationRequests
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationRequestDto'
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
...
    AccountStatus:
      type: string
      enum:
      - UNVERIFIED
      - PENDING
      - ACTIVE
      - INACTIVE
      - DELETED
      - BANNED
      - REJECTED
...
    RegistrationRequestDto:
      required:
      - accountStatus
      - accountType
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - registrationDate
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        avatarImageUrl:
          type: string
        phoneNumber:
          type: string
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
        rejectionReason:
          type: string
        registrationDate:
          type: string
          format: date-time
        accountType:
          type: string
          enum:
          - ADMIN
          - CLIENT
          - FREELANCER